requirements.txt
asgiref==3.8.1
boto3==1.36.11
botocore==1.36.11
django==5.1.5
django-cors-headers==4.6.0
django-environ==0.11.2
django-inscode==0.1.14
django-soft-delete==1.0.18
django-storages==1.14.4
factory-boy==3.3.3
faker==35.2.0
jmespath==1.0.1
marshmallow==3.26.1
packaging==24.2
psycopg==3.2.4
python-dateutil==2.9.0.post0
pytz==2024.2
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
validate-docbr==1.10.0
